add missing deprecation in yml. Move AttachmentInfo command and handlers to Attachment namespace
