+++ src/Adapter/Product/QueryHandler/GetProductForEditingHandler.php
@@ -36,7 +36,7 @@ use PrestaShop\PrestaShop\Adapter\Product\Stock\Repository\StockAvailableReposit
 use PrestaShop\PrestaShop\Adapter\Product\VirtualProduct\Repository\VirtualProductFileRepository;
 use PrestaShop\PrestaShop\Adapter\Tax\TaxComputer;
 use PrestaShop\PrestaShop\Core\Domain\Country\ValueObject\CountryId;
-use PrestaShop\PrestaShop\Core\Domain\Product\Attachment\QueryResult\ProductAttachmentInfo;
+use PrestaShop\PrestaShop\Core\Domain\Attachment\QueryResult\AttachmentInfo;
 use PrestaShop\PrestaShop\Core\Domain\Product\ProductCustomizabilitySettings;
 use PrestaShop\PrestaShop\Core\Domain\Product\Query\GetProductForEditing;
 use PrestaShop\PrestaShop\Core\Domain\Product\QueryHandler\GetProductForEditingHandlerInterface;
@@ -155,7 +155,7 @@ final class GetProductForEditingHandler implements GetProductForEditingHandlerIn
     /**
      * @param ProductId $productId
      *
-     * @return ProductAttachmentInfo[]
+     * @return AttachmentInfo[]
      */
     private function getAttachments(ProductId $productId): array
     {
@@ -163,7 +163,7 @@ final class GetProductForEditingHandler implements GetProductForEditingHandlerIn
 
         $productAttachments = [];
         foreach ($attachments as $attachment) {
-            $productAttachments[] = new ProductAttachmentInfo(
+            $productAttachments[] = new AttachmentInfo(
                 (int) $attachment['id_attachment'],
                 $attachment['name'],
                 $attachment['file_name'],
