+++ src/Core/Domain/Attachment/QueryHandler/GetAttachmentInfoHandlerInterface.php
@@ -23,32 +23,21 @@
  * @copyright Since 2007 PrestaShop SA and Contributors
  * @license   https://opensource.org/licenses/OSL-3.0 Open Software License (OSL 3.0)
  */
-declare(strict_types=1);
 
-namespace PrestaShop\PrestaShop\Core\Domain\Product\Attachment\Query;
+namespace PrestaShop\PrestaShop\Core\Domain\Attachment\QueryHandler;
 
-use PrestaShop\PrestaShop\Core\Domain\Attachment\ValueObject\AttachmentId;
+use PrestaShop\PrestaShop\Core\Domain\Attachment\Query\GetAttachmentInfo;
+use PrestaShop\PrestaShop\Core\Domain\Attachment\QueryResult\AttachmentInfo;
 
-class GetProductAttachmentInfo
+/**
+ * Defines contract to handle @see GetAttachmentInfo query
+ */
+interface GetAttachmentInfoHandlerInterface
 {
     /**
-     * @var AttachmentId
+     * @param GetAttachmentInfo $query
+     *
+     * @return AttachmentInfo
      */
-    private $attachmentId;
-
-    /**
-     * @param int $attachmentId
-     */
-    public function __construct(int $attachmentId)
-    {
-        $this->attachmentId = new AttachmentId($attachmentId);
-    }
-
-    /**
-     * @return AttachmentId
-     */
-    public function getAttachmentId(): AttachmentId
-    {
-        return $this->attachmentId;
-    }
+    public function handle(GetAttachmentInfo $query): AttachmentInfo;
 }
