+++ src/PrestaShopBundle/Resources/config/services/adapter/product.yml
@@ -291,14 +291,6 @@ services:
             - name: tactician.handler
               command: PrestaShop\PrestaShop\Core\Domain\Product\Command\RemoveAllAssociatedProductAttachmentsCommand
 
-    prestashop.adapter.query_handler.get_product_attachment_info_handler:
-        class: PrestaShop\PrestaShop\Adapter\Product\QueryHandler\GetProductAttachmentInfoHandler
-        arguments:
-            - '@prestashop.adapter.attachment.attachment_repository'
-        tags:
-            - name: tactician.handler
-              command: PrestaShop\PrestaShop\Core\Domain\Product\Attachment\Query\GetProductAttachmentInfo
-
     prestashop.adapter.product.validate.product_supplier_validator:
         class: PrestaShop\PrestaShop\Adapter\Product\Validate\ProductSupplierValidator
         arguments:
